Update App.jsx
The SplitBill component preserves the states from the previous render when a user  was selected. 
That’s because SplitBill component is not unique across the renders. React cannot differentiate the component instances when key prop is not provided, instead it uses the same instance with same states from the previous render to the next render. the solution is to provide the key prop which differentiates SplitBill component across the renders when the user is switched.